MBS-11550: Move PostParameters to static/scripts for translation
Hydrated components don't get translated unless they are
under static/scripts, apparently.
