+++ root/static/scripts/confirm-seed.js
@@ -7,5 +7,5 @@
  * later version: http://www.gnu.org/licenses/gpl-2.0.txt
  */
 
-import '../../components/PostParameters';
+import './common/components/PostParameters';
 import '../../main/components/ConfirmSeedButtons';
