+++ root/static/styles/forms.less
@@ -826,7 +826,7 @@ button.guess-timezone {
     }
 }
 
-/* root/components/PostParameters.js */
+/* root/static/scripts/common/components/PostParameters.js */
 
 .post-parameters {
     a.expand-link {
